% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beads_GR_to_UM.R
\name{beads_GR_to_UM}
\alias{beads_GR_to_UM}
\title{Transforms nBeads from probes to CpGs}
\usage{
beads_GR_to_UM(nBeads, rgSet)
}
\arguments{
\item{nBeads}{A matrix containing the number of beads (probes as rows, samples as columns)}

\item{rgSet}{An rgSet object imported by minfi (see minfi::read.metharray.exp for details)}
}
\value{
\code{nBeads_cg}: a matrix containing number of beads (CpGs as rows, samples as columns)
}
\description{
Transforms the matrix of number of beads from Green/Red to Unmethylated/Methylated channels
}
\details{
Illumina DNA methylation microarrays detection rely on three types of probes:
\itemize{
  \item type-I Red: 2 probes/addresses per CpG, informative on the Red channel exclusively.
  \item type-I Green: 2 probes/addresses per CpG, informative on the Green channel exclusively.
  \item type-II: 1 probe/address per CpG, informative in both channels.
}
Due to the beadchip technology, each probe has been assigned to a random number of beads.
beads_GR_to_UM converts the number of beads per probe and per sample to
the number of beads per CpG and per sample.
This requires a criteria for type-I probes for which 2 probes target the same CpG. In this case,
the minimum number between the two probes targetting a type-I probe is chosen.
}
\examples{
rgSet = read.metharray.exp(getwd(), extended = TRUE)
nBeads = assay(rgSet, "NBeads")   # Number of Beads across probes
beads_GR_to_UM(nBeads, rgSet)
}
