% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bimodality_coefficient.R
\name{bimodality_coefficient}
\alias{bimodality_coefficient}
\title{Bimodality coefficient (Copied from the deprecated R-package modes)}
\usage{
bimodality_coefficient(x, finite = TRUE)
}
\arguments{
\item{x}{Data vector.}

\item{finite}{Should the finite sample size correction be applied to the skewness and kurtosis measures? Defaults to TRUE.}
}
\value{
A bimodality coefficient (numeric)
}
\description{
Computes bimodality coefficient.
}
\details{
Due to discontinuation from CRAN (Package ‘modes’ was removed from the CRAN repository; Archived on 2020-03-03 as check
problems were not corrected despite reminders), we have copied the required functions into UMtools to ease installation (and slightly modified).
The original documentation states the following: "This function calculates the Bimodality Coefficient of a data vector with the option
for a finite sample (bias) correction. This bias correction is important to correct for the (well-documented) finite sample bias.
The bimodality coefficient has a range of zero to one (that is: [0,1]) where a value greater than "5/9" suggests bimodality.
The maximum value of one ("1") can only be reached when the distribution is composed of two point masses".
}
\examples{
data <- c(rnorm(15, 0, 1), rnorm(21, 5, 1))
hist(data)
bimodality_coefficient(data, TRUE)
}
