% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cv.R
\name{compute_cv}
\alias{compute_cv}
\title{Computes coefficient of variation of the log of the total intensity signal}
\usage{
compute_cv(M_SD, U_SD, M, U, alpha1 = 100, alpha2 = 1)
}
\arguments{
\item{M_SD}{Methylated standard deviation intensity matrix (CpGs as rows, samples as columns)}

\item{U_SD}{Unmethylated fluorescence standard deviation intensity matrix (CpGs as rows, samples as columns)}

\item{M}{Methylated fluorescence mean intensity matrix (CpGs as rows, samples as columns)}

\item{U}{Unmethylated fluorescence mean intensity matrix (CpGs as rows, samples as columns)}
}
\value{
A matrix \code{CV} (CpGs as rows, samples as columns)
}
\description{
a
}
\details{
a
}
\examples{
GR_to_UM(Red, Grn, rgSet)
}
