% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR_probes-data.R
\docType{data}
\name{CR_probes}
\alias{CR_probes}
\title{Cross-reactive probes}
\format{
An object of class
}
\source{
a
}
\usage{
data(CR_probes)
}
\description{
Cross-reactive probes
}
\examples{
data(CR_probes)
}
\references{
a
}
\keyword{datasets}
