% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_jitter_plot.R
\name{density_jitter_plot}
\alias{density_jitter_plot}
\title{Plots density jitter plot}
\usage{
density_jitter_plot(mat, CpG, sex = NULL, alpha = 0.5, lambda = 0.05)
}
\arguments{
\item{Red}{A red fluorescence intensity matrix (probes as rows, samples as columns)}

\item{Grn}{A green fluorescence intensity matrix (probes as rows, samples as columns)}

\item{rgSet}{An rgSet object imported by minfi (see minfi::read.metharray.exp for details)}
}
\value{
A list including matrices \code{M} and \code{U} (CpGs as rows, samples as columns)
}
\description{
a
}
\details{
a
}
\examples{
GR_to_UM(Red, Grn, rgSet)
}
