% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_bigmat.R
\name{export_bigmat}
\alias{export_bigmat}
\title{Export big matrices}
\usage{
export_bigmat(bigmat, filename, nThread = 1)
}
\arguments{
\item{bigmat}{A matrix (probes as rows, samples as columns)}

\item{filename}{Name of the output file. The date will be appended automatically.}

\item{nThread}{Number of CPU cores to employ}
}
\description{
Exports large matrices with the help of data.table::fwrite
}
\examples{
export_bigmat(M, "M.txt")
}
