% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_to_UM.R
\name{GR_to_UM}
\alias{GR_to_UM}
\title{Transforms G/R fluorescence signals into U/M}
\usage{
GR_to_UM(Red, Grn, rgSet)
}
\arguments{
\item{Red}{A red fluorescence intensity matrix (probes as rows, samples as columns)}

\item{Grn}{A green fluorescence intensity matrix (probes as rows, samples as columns)}

\item{rgSet}{An rgSet object imported by minfi (see minfi::read.metharray.exp for details)}
}
\value{
A list including matrices \code{M} and \code{U} (CpGs as rows, samples as columns)
}
\description{
Transforms G/R fluorescence signals into U/M
}
\examples{
GR_to_UM(Red, Grn, rgSet)
}
