% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k_per_CpG.R
\name{k_per_CpG}
\alias{k_per_CpG}
\title{Computes bimodality per CpG across samples in a coefficient of variation matrix}
\usage{
k_per_CpG(df, minPts = 25, reach = seq(0.328, 0.352, 0.004), R = 2)
}
\arguments{
\item{CV}{A coefficient of variation matrix (probes as rows, samples as columns)}
}
\value{
A vector of coefficient of bimodality per CpG, \code{BC}
}
\description{
a
}
\details{
a
}
\examples{
compute_BC_CV(CV)
}
