% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kcall_CpG.R
\name{Kcall_CpG}
\alias{Kcall_CpG}
\title{Transforms number of beads matrix from G/R to U/M}
\usage{
Kcall_CpG(
  CpG,
  M,
  U,
  minPts = 25,
  plot = T,
  reach = seq(0.328, 0.352, 0.004),
  R = 2
)
}
\arguments{
\item{nBeads}{A matrix containing the number of beads per probe and per sample (probes as rows, samples as columns)}

\item{rgSet}{An rgSet object imported by minfi (see minfi::read.metharray.exp for details)}
}
\value{
A matrix containing number of beads per CpG and per sample, \code{nBeads_cg}, (CpGs as rows, samples as columns)
}
\description{
a
}
\details{
a
}
\examples{
beads_GR_to_UM(nBeads, rgSet)
}
