% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_EW_Kcalling.R
\name{par_EW_Kcalling}
\alias{par_EW_Kcalling}
\title{Epigenome-wide K-calling}
\usage{
par_EW_Kcalling(M, U, minPts = 12, eps = 0.035, nThread = NULL)
}
\arguments{
\item{M}{Methylated fluorescence mean intensity matrix (CpGs as rows, samples as columns)}

\item{U}{Unmethylated fluorescence mean intensity matrix (CpGs as rows, samples as columns)}

\item{minPts}{dbscan parameter. Minimum numbers of points in the eps region to consider a core point; help(dbscan, dbscan) for more details.}

\item{eps}{dbscan parameter. Size of the neighbourhood; help(dbscan, dbscan) for more details.}

\item{nThread}{Number of CPU cores to employ}
}
\value{
Vector. Predicted number of clusters for each CpG.
}
\description{
Same as Kcall_CpG but it performs K-calling on all the CpGs (rows) of a given matrix.
Also, it allows parallel processing. If nThread is not specified, it will automatically detect
how many clusters are available in the machine and will use N-1 cores.
}
\examples{
annotation <- getAnnotation(rgSet)
chrY = rownames(annotation)[annotation$chr == "chrY"]
K_vec = par_EW_Kcalling(M[chrY,], U[chrY,])

}
