% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.R
\name{skewness}
\alias{skewness}
\title{Skewness (Copied from the deprecated R-package modes)}
\usage{
skewness(x, finite = TRUE)
}
\arguments{
\item{x}{Data vector.}

\item{finite}{Should the finite sample size correction be applied? Defaults to TRUE.}
}
\value{
Skewness (numeric)
}
\description{
Computes skewness.
}
\details{
Due to discontinuation from CRAN (Package ‘modes’ was removed from the CRAN repository; Archived on 2020-03-03 as check
problems were not corrected despite reminders), we have copied the required functions into UMtools to ease installation (and slightly modified).
The original documentation states the following: "This function calculates the skewness of a data with optional bias correction.
The skewness is a measure of the symmetry of a distrbution. A negative skewness means the data is left skewed or has a fat left tail.
The converse is true for a positive skew".
}
\examples{
data <- c(rnorm(15, 0, 1), rnorm(21, 5, 1))
hist(data)
skewness(data, TRUE)
}
