% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_k_caller.R
\name{train_k_caller}
\alias{train_k_caller}
\title{K-caller performance for a given set of eps and minPts}
\usage{
train_k_caller(M, U, training_set, minPts, eps, nThread = NULL)
}
\arguments{
\item{M}{Methylated fluorescence mean intensity matrix (CpGs as rows, samples as columns)}

\item{U}{Unmethylated fluorescence mean intensity matrix (CpGs as rows, samples as columns)}

\item{training_set}{List. As an example, see data(training_set). More details in help(training_set)}

\item{minPts}{dbscan parameter. Minimum numbers of points in the eps region to consider a core point; help(dbscan, dbscan) for more details.}

\item{eps}{dbscan parameter. Size of the neighbourhood; help(dbscan, dbscan) for more details.}

\item{nThread}{Number of CPU cores to employ}
}
\value{
List. Containing confusion matrix and macro-Precission, macro-Recall and Macro F1-score for
1-3 clusters and 1-4 clusters. Given the rarity of K=4 clusters, it is not fair to give the same weight
in evaluation of the multi-class classification performance. To decide what combination of parameters to employ,
we recommend to use macro F1-score for 1-3 clusters.
}
\description{
It allows the quantification of the K-caller's performance for a given dataset, training set of probes
and parameters eps and minPts.
}
\examples{
data(training_set)
train_k_caller(M, U, training_set, 3, 0.07)

}
