% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UM_plot.R
\name{UM_plot}
\alias{UM_plot}
\title{Transforms G/R fluorescence signals into U/M}
\usage{
UM_plot(M, U, CpG, sex = NULL, xlim = NULL, ylim = NULL, alpha = 0.5)
}
\arguments{
\item{Red}{A red fluorescence intensity matrix (probes as rows, samples as columns)}

\item{Grn}{A green fluorescence intensity matrix (probes as rows, samples as columns)}

\item{rgSet}{An rgSet object imported by minfi (see minfi::read.metharray.exp for details)}
}
\value{
A list including matrices \code{M} and \code{U} (CpGs as rows, samples as columns)
}
\description{
a
}
\details{
a
}
\examples{
GR_to_UM(Red, Grn, rgSet)
}
