% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_cometh.R
\name{Visualize_cometh}
\alias{Visualize_cometh}
\title{Co-methylation plot}
\usage{
Visualize_cometh(
  annotation,
  CpG,
  distance,
  L_bound,
  R_bound,
  beta_mat,
  probe2exclude = NULL,
  max_y = 11
)
}
\arguments{
\item{annotation}{A DNA methylation microarray annotation object. For example, getAnnotation(IlluminaHumanMethylation450kanno.ilmn12.hg19)}

\item{CpG}{A CpG identifier (e.g. "cg15771735")}

\item{distance}{A distance that defines the neighbourhood explored for co-methylation (e.g. pos ± distance)}

\item{L_bound}{It allows to easily extend the number of CpGs on the left flank}

\item{R_bound}{It allows to easily extend the number of CpGs on the right flank}

\item{beta_mat}{A numeric matrix (CpGs as rows, samples as columns). We recommend to use M-values, though the function works
fine with beta-values.}

\item{probe2exclude}{A list of potentially artefactual probes. These will be coloured in purple on the comethylation plot.}

\item{max_y}{It regulates the height of the co-methylation triangle. For more details, see help(plotHic, Sushi)}
}
\value{
Graphics + vector. A vector of cg ids visualized in the co-methylation plot is returned.
}
\description{
It visualizes co-methylation of a target probe with its neighbours with a HiC-like plot.
}
\details{
Comethylation squared within the interval [0, 1] is visualized. Distance between probes \eqn{\Delta\delta} is also available on the plot.
}
\examples{
res = Visualize_cometh(annotation = annotation, CpG = 'cg14911689', distance = 1000,
                       L_bound = 3, R_bound = 2, beta_mat = beta_value, max_y = 5)
}
