% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bGMM.R
\name{bGMM}
\alias{bGMM}
\title{Bivariate Gaussian Mixture Model}
\usage{
bGMM(
  M,
  U,
  CpG,
  K,
  stable.solution = TRUE,
  min.n = NULL,
  min.n.iter = 2000,
  method = "em.EM",
  EMC = .EMC
)
}
\arguments{
\item{M}{A matrix containing the methylated intensities (CpGs as rows, samples as columns)}

\item{U}{A matrix containing the unmethylated intensities (CpGs as rows, samples as columns)}

\item{CpG}{A CpG identifier (e.g. "cg15771735")}

\item{K}{Targetted number of clusters}

\item{...}{Control functions for EMCluster}
}
\value{
A vector of classes in the same order as the columns
}
\description{
Fits a Gaussian Mixture Model (bGMM) with K components on the Unmethylated/Methylated plane
}
\details{
It employs the function init.EM from the R-package EMCluster that fits a GMM employing the expectation
maximization algorithm. Run help(initEM) for more information.
}
\examples{
rgSet = read.metharray.exp(getwd(), extended = TRUE)
Grn = assay(rgSet, "Green")       # Green mean across beads
Red = assay(rgSet, "Red")         # Red mean across beads
M_U = GR_to_UM(Red, Grn, rgSet)
bGMM(M_U$M, M_U$U, "cg00814218", 3)
}
