% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BC_CV.R
\name{compute_BC_CV}
\alias{compute_BC_CV}
\title{Computes bimodality per CpG across samples in a coefficient of variation matrix}
\usage{
compute_BC_CV(CV)
}
\arguments{
\item{CV}{A coefficient of variation matrix (probes as rows, samples as columns)}
}
\value{
A vector of coefficient of bimodality per CpG, \code{BC}
}
\description{
a
}
\details{
a
}
\examples{
compute_BC_CV(CV)
}
