% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_jitter_plot.R
\name{density_jitter_plot}
\alias{density_jitter_plot}
\title{Density-jitter plot}
\usage{
density_jitter_plot(
  mat,
  CpG,
  sex = NULL,
  alpha = 0.5,
  lambda = 0.05,
  main = NULL
)
}
\arguments{
\item{mat}{Matrix of a continous varible such as beta-value, M-value, CVlogT, etc (CpGs as rows and samples as columns)}

\item{CpG}{Targeted CpG}

\item{sex}{vector of sex. First level is expected to be Female (coloured pink). Check levels(sex) to see if that is the case.}

\item{alpha}{Transparency}

\item{lambda}{Regulates the random spread of the points}

\item{main}{Title name}
}
\value{
Graphics
}
\description{
Overlaps a density plot with a jitter plot. A random y-component is added to cause the spread of
points.
}
\examples{
density_jitter_plot(beta_value, "cg00050873", pheno$sex)
}
