% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_bigmat.R
\name{export_bigmat}
\alias{export_bigmat}
\title{Transforms number of beads matrix from G/R to U/M}
\usage{
export_bigmat(bigmat, filename, nThread = 1)
}
\arguments{
\item{nBeads}{A matrix containing the number of beads per probe and per sample (probes as rows, samples as columns)}

\item{rgSet}{An rgSet object imported by minfi (see minfi::read.metharray.exp for details)}
}
\value{
A matrix containing number of beads per CpG and per sample, \code{nBeads_cg}, (CpGs as rows, samples as columns)
}
\description{
a
}
\details{
a
}
\examples{
beads_GR_to_UM(nBeads, rgSet)
}
