% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_to_UM.R
\name{GR_to_UM}
\alias{GR_to_UM}
\title{GR_to_UM}
\usage{
GR_to_UM(Red = NULL, Grn = NULL, rgSet, what = NULL, nBeads = NULL)
}
\arguments{
\item{Red}{Mean or SD Red fluorescence intensity matrix (probes as rows, samples as columns)}

\item{Grn}{Mean or SD Green fluorescence intensity matrix (probes as rows, samples as columns)}

\item{rgSet}{An rgSet object imported by minfi (see minfi::read.metharray.exp for details)}

\item{what}{Either 'Mean', 'SD' or 'NBeads'}

\item{nBeads}{A matrix containing the number of beads (probes as rows, samples as columns)}
}
\value{
A list including matrices \emph{M} and \emph{U} (CpGs as rows, samples as columns) when \emph{what} is "Mean"
or "SD" and or a matrix \emph{nBeads}, containing number of beads (CpGs as rows, samples as columns) when
\emph{what} is equal to "NBeads".
}
\description{
Transforms Green/Red fluorescence signals into Unmethylated/Methylated channels
}
\details{
Illumina DNA methylation microarrays detection rely on three types of probes:
\itemize{
\item type-I Red: 2 probes/addresses per CpG, informative on the Red channel exclusively.
\item type-I Green: 2 probes/addresses per CpG, informative on the Green channel exclusively.
\item type-II: 1 probe/address per CpG, informative in both Green and Red channels.
}

When \emph{what} is equal to "SD" or "Mean", Methylated and unmethylated intensities are computed as followed:\tabular{lcr}{
    \tab \strong{Methylated} \tab \strong{Unmethylated} \cr
   \strong{Type-II} \tab Green (addressA) \tab Red (addressA) \cr
   \strong{Type-I Red} \tab Red (addressB) \tab Red (addressA) \cr
   \strong{Type-I Green} \tab Green (addressB) \tab Green (addressA) \cr
}


Also, due to the beadchip technology, each probe has been assigned to a random number of beads.
If \emph{what} == "NBeads", it converts the number of beads per probe and per sample to
the number of beads per CpG and per sample. This requires a criteria for type-I probes for
which 2 probes target the same CpG. In this case,the minimum number between the two probes
targetting a type-I probe is chosen.
}
\examples{
# Read IDAT files. Without the extended option, SD cannot be acquired
rgSet <- read.metharray.exp(getwd(), extended = TRUE)

# Green/Red mean to Unmethylated/Methylated mean intensities
Grn <- assay(rgSet, "Green")
Red <- assay(rgSet, "Red")
UM <- GR_to_UM(Red = Red, Grn = Grn, rgSet = rgSet, what = "Mean")

# Green/Red SD to Unmethylated/Methylated intensity SDs
GrnSD <- assay(rgSet, "GreenSD")
RedSD <-assay(rgSet, "RedSD")
UM_SD <- GR_to_UM(Red = RedSD, Grn = GrnSD, rgSet = rgSet, what = "SD")

# nBeads in probes to nBeads per CpG
nBeads <- assay(rgSet, "NBeads")
nBeads <- GR_to_UM(nBeads = nBeads, rgSet = rgSet, what = "NBeads")
}
