% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR_to_UM.R
\name{GR_to_UM}
\alias{GR_to_UM}
\title{Transforms Green/Red fluorescence signals into Unmethylated/Methylated channels}
\usage{
GR_to_UM(Red, Grn, rgSet)
}
\arguments{
\item{Red}{A red fluorescence intensity matrix (probes as rows, samples as columns)}

\item{Grn}{A green fluorescence intensity matrix (probes as rows, samples as columns)}

\item{rgSet}{An rgSet object imported by minfi (see minfi::read.metharray.exp for details)}
}
\value{
A list including matrices \code{M} and \code{U} (CpGs as rows, samples as columns)
}
\description{
Transforms
}
\details{
Illumina DNA methylation microarrays detection rely on three types of probes:
\itemize{
\item type-I Red: 2 probes/addresses per CpG, informative on the Red channel exclusively.
\item type-I Green: 2 probes/addresses per CpG, informative on the Green channel exclusively.
\item type-II: 1 probe/address per CpG, informative in both Green and Red channels.
}

Methylated and unmethylated intensities are computed as followed:\tabular{lcr}{
    \tab \strong{Methylated} \tab \strong{Unmethylated} \cr
   \strong{Type-II} \tab Green (addressA) \tab Red (addressA) \cr
   \strong{Type-I Red} \tab Red (addressB) \tab Red (addressA) \cr
   \strong{Type-I Green} \tab Green (addressB) \tab Green (addressA) \cr
}
}
\examples{
rgSet = read.metharray.exp(getwd(), extended = TRUE)
Grn = assay(rgSet, "Green")       # Green mean across beads
Red = assay(rgSet, "Red")         # Red mean across beads
GR_to_UM(Red, Grn, rgSet)
GrnSD = assay(rgSet, "GreenSD")   # Green SD across beads
RedSD = assay(rgSet, "RedSD")     # Red SD across beads
GR_to_UM(RedSD, GrnSD, rgSet)
}
