% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_EW_Kcalling.R
\name{par_EW_Kcalling}
\alias{par_EW_Kcalling}
\title{Computes bimodality per CpG across samples in a coefficient of variation matrix}
\usage{
par_EW_Kcalling(M, U, minPts = 25, reach = seq(0.328, 0.352, 0.004), R = 2)
}
\arguments{
\item{CV}{A coefficient of variation matrix (probes as rows, samples as columns)}
}
\value{
A vector of coefficient of bimodality per CpG, \code{BC}
}
\description{
a
}
\details{
a
}
\examples{
compute_BC_CV(CV)
}
