% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_cometh.R
\name{Visualize_cometh}
\alias{Visualize_cometh}
\title{Computes bimodality per CpG across samples in a coefficient of variation matrix}
\usage{
Visualize_cometh(
  annotation,
  CpG,
  distance,
  L_bound = 0,
  R_bound = 0,
  beta_mat,
  probe2flag = NULL,
  max_y = 11,
  cgHeightLabel = -1,
  deltaposHeightLabel = -0.2,
  chrHeightLabel = -2
)
}
\arguments{
\item{CV}{A coefficient of variation matrix (probes as rows, samples as columns)}
}
\value{
A vector of coefficient of bimodality per CpG, \code{BC}
}
\description{
a
}
\details{
a
}
\examples{
compute_BC_CV(CV)
}
